Fix Interval/UTC time missing and prevent bad sample from being read
This fix primarily addresses issues where Interval/UTC columns weren't
showing up in the log files due to a design flaw in how to we decide to
print those columns and the multiplication propteries of our sample rates.

This patch also adds a sanity check to our LogMessage to ensure that it is
not more than 10ms old.  Else we can't gaurantee the data is good and should
just not use that packet.